Peticiones get y delete funcioando

+++ demo/src/main/java/com/innovatech/demo/Repository/EventRepository.java
@@ -10,7 +10,7 @@ import com.innovatech.demo.Entity.EventEntity;
 @Repository
 public interface EventRepository extends JpaRepository<EventEntity, Long> {
 
-    Optional<EventEntity> findById(Long id);
+    Optional<EventEntity> findById(int id);
 
     EventEntity findByName(String name);
 }
